Fixed Intersection.h

+++ src/Intersection.h
@@ -1,4 +1,4 @@
 #ifndef INTERSECTION_H
 #define INTERSECTION_H
-int Intersection(double x1, double y1, double x2, double y2, float r1, float r2);
+int Intersection(int x1, int y1, int x2, int y2, float r1, float r2);
 #endif // INTERSECTION_H
